UI-346 Misc refactors
