+++ de-lib/src/main/java/org/iplantc/de/admin/desktop/client/communities/AdminCommunitiesView.java
@@ -24,6 +24,8 @@ import java.util.List;
 
 /**
  * @author aramsey
+ *
+ * A view that enables DE admins to manage communities in the DE and tag apps with communities
  */
 public interface AdminCommunitiesView extends IsWidget,
                                               HasHandlers,
@@ -104,34 +106,92 @@ public interface AdminCommunitiesView extends IsWidget,
                                 EditCommunityClicked.EditCommunityClickedHandler,
                                 CategorizeButtonClicked.CategorizeButtonClickedHandler,
                                 DeleteCommunityClicked.DeleteCommunityClickedHandler {
+        /**
+         * Initializes the view and adds the view to the specified container
+         * @param container
+         */
         void go(HasOneWidget container);
 
+        /**
+         * @return the view
+         */
         AdminCommunitiesView getView();
 
+        /**
+         * @param el
+         * @return the community that is associated with the element
+         */
         Group getCommunityFromElement(Element el);
 
+        /**
+         * @return the community that is currently selected
+         */
         Group getSelectedCommunity();
 
+        /**
+         * Sets the base static ID for the view
+         * @param id
+         */
         void setViewDebugId(String id);
 
+        /**
+         * This method is called when a user drags and drops apps to a community in order
+         * to tag those apps with that community
+         * @param apps
+         * @param community
+         */
         void appsDNDtoCommunity(List<App> apps, Group community);
 
+        /**
+         * This method is called when a user drags and drops a community to an app in
+         * order to tag the app with that community
+         * @param community
+         * @param targetApp
+         */
         void communityDNDtoApp(Group community, App targetApp);
     }
 
+    /**
+     *
+     * @param el
+     * @return the community that is associated with the element
+     */
     Group getCommunityFromElement(Element el);
 
+    /**
+     *
+     * @return the currently selected community
+     */
     Group getSelectedCommunity();
 
+    /**
+     * Display a panel indicating to the user that there are currently no communities
+     */
     void showNoCommunitiesPanel();
 
+    /**
+     * Display a panel that will present the user with the communities list
+     */
     void showCommunitiesPanel();
 
+    /**
+     * Deselects all hierarchies
+     */
     void deselectHierarchies();
 
+    /**
+     * Selects the specified community in the view, typically to load the apps in that community
+     * @param community
+     */
     void selectCommunity(Group community);
 
+    /**
+     * Masks the communities tree, typically while data is loading
+     */
     void maskCommunities();
 
+    /**
+     * Unmasks the communities tree
+     */
     void unmaskCommunities();
 }
