+++ de-lib/src/main/java/org/iplantc/de/admin/desktop/client/communities/gin/AdminCommunitiesGinModule.java
@@ -1,14 +1,14 @@
 package org.iplantc.de.admin.desktop.client.communities.gin;
 
 import org.iplantc.de.admin.desktop.client.communities.AdminCommunitiesView;
+import org.iplantc.de.admin.desktop.client.communities.ManageCommunitiesView;
 import org.iplantc.de.admin.desktop.client.communities.presenter.AdminCommunitiesPresenterImpl;
 import org.iplantc.de.admin.desktop.client.communities.presenter.AdminManageCommunitiesPresenterImpl;
 import org.iplantc.de.admin.desktop.client.communities.service.AdminCommunityServiceFacade;
 import org.iplantc.de.admin.desktop.client.communities.service.impl.AdminCommunityServiceFacadeImpl;
 import org.iplantc.de.admin.desktop.client.communities.views.AdminCommunitiesViewImpl;
-import org.iplantc.de.apps.client.ManageCommunitiesView;
+import org.iplantc.de.admin.desktop.client.communities.views.ManageCommunitiesViewImpl;
 import org.iplantc.de.apps.client.gin.CommunityTreeStoreProvider;
-import org.iplantc.de.apps.client.views.communities.ManageCommunitiesViewImpl;
 import org.iplantc.de.client.models.groups.Group;
 
 import com.google.gwt.inject.client.AbstractGinModule;
