+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/admin/communities/CommunityDisplayStrings.java
@@ -44,4 +44,6 @@ public interface CommunityDisplayStrings extends Messages{
     String communityDeleted(String name);
 
     String failedToAddCommunityAdmin(String adminName, String communityName);
+
+    String searchEmptyText();
 }
