+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/admin/communities/CommunityDisplayStrings.properties
@@ -15,4 +15,5 @@ externalAppDND=External app(s): {0}
 failedToAddCommunityAdmin=Failed to add {0} as an admin to {1}
 hierarchyPreviewHeader=DE Hierarchy Preview
 hierarchyTreePanel=Hierarchies
+searchEmptyText=Search Apps...
 selectCommunitiesFor=Select communities for {0}
