+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/admin/communities/ManageCommunitiesViewDefaultAppearance.java
@@ -1,6 +1,6 @@
-package org.iplantc.de.theme.base.client.apps.communities;
+package org.iplantc.de.theme.base.client.admin.communities;
 
-import org.iplantc.de.apps.client.ManageCommunitiesView;
+import org.iplantc.de.admin.desktop.client.communities.ManageCommunitiesView;
 import org.iplantc.de.client.models.collaborators.Subject;
 import org.iplantc.de.client.models.groups.Group;
 import org.iplantc.de.resources.client.IplantResources;
