+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/apps/Applications.gwt.xml
@@ -33,10 +33,6 @@
         <when-type-is class="org.iplantc.de.apps.client.CommunitiesView.Appearance"/>
     </replace-with>
 
-    <replace-with class="org.iplantc.de.theme.base.client.apps.communities.ManageCommunitiesViewDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.ManageCommunitiesView.Appearance"/>
-    </replace-with>
-
     <!-- Cells -->
     <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppCommentCellDefaultAppearance">
         <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppCommentCell.AppCommentCellAppearance"/>
